Create curve_point file
